Simplify otel attributes file structure
