+++ lambda-otel4s/shared/src/main/scala/feral/lambda/otel4s/LambdaContextAttributes.scala
@@ -20,36 +20,8 @@ import feral.lambda.Context
 import org.typelevel.otel4s.AttributeKey
 import org.typelevel.otel4s.Attributes
 
-/**
- * Temporary aliases for Lambda platform attributes in otel4s-semconv-experimental
- */
-private[otel4s] object LambdaContextAttributes {
-  val FaasInvocationId = AttributeKey.string("faas.invocation_id")
-  val FaasTrigger = AttributeKey.string("faas.trigger")
-  object FaasTriggerValue {
-    object Pubsub {
-      val value = "pubsub"
-    }
-    object Datasource {
-      val value = "datasource"
-    }
-    object Http {
-      val value = "http"
-    }
-  }
-
-  // ARN
-  val CloudResourceId = AttributeKey.string("cloud.resource_id")
-  val FaasInstance = AttributeKey.string("faas.instance")
-  val FaasMaxMemory = AttributeKey.long("faas.max_memory")
-  val FaasName = AttributeKey.string("faas.name")
-  val FaasVersion = AttributeKey.string("faas.version")
-  val CloudProvider = AttributeKey.string("cloud.provider")
-  object CloudProviderValue {
-    object Aws {
-      val value = "aws"
-    }
-  }
+object LambdaContextAttributes {
+  import OtelAttributes._
 
   def apply[F[_]](context: Context[F]): Attributes = {
     Attributes(
