+++ lambda-otel4s/shared/src/test/scala/feral/lambda/otel4s/OtelAttributesTest.scala
@@ -19,15 +19,17 @@ package feral.lambda.otel4s
 import munit.FunSuite
 import org.typelevel.otel4s.semconv.experimental.attributes.CloudExperimentalAttributes
 import org.typelevel.otel4s.semconv.experimental.attributes.FaasExperimentalAttributes
+import org.typelevel.otel4s.semconv.experimental.attributes.MessagingExperimentalAttributes
 
 /**
  * Check attributes match experimental ones
  */
-class LambdaContextAttributesTest extends FunSuite {
+class OtelAttributesTest extends FunSuite {
 
-  import LambdaContextAttributes._
+  import OtelAttributes._
   val otel4s = FaasExperimentalAttributes
   val otel4sCloud = CloudExperimentalAttributes
+  val otel4sMessaging = MessagingExperimentalAttributes
 
   test("FaasInvocationId") {
     val value = "value"
@@ -108,4 +110,45 @@ class LambdaContextAttributesTest extends FunSuite {
     val otel = otel4sCloud.CloudProviderValue.Aws.value
     assertEquals(ours, otel)
   }
+
+  test("MessagingSystem") {
+    val value = "value"
+    val ours = MessagingSystem(value)
+    val otel = otel4sMessaging.MessagingSystem(value)
+    assertEquals(ours, otel)
+  }
+
+  test("MessagingSystemValue.AwsSqs") {
+    val ours = MessagingSystemValue.AwsSqs.value
+    val otel = otel4sMessaging.MessagingSystemValue.AwsSqs.value
+    assertEquals(ours, otel)
+  }
+
+  test("MessagingOperationType") {
+    val value = "value"
+    val ours = MessagingOperationType(value)
+    val otel = otel4sMessaging.MessagingOperationType(value)
+    assertEquals(ours, otel)
+  }
+
+  test("MessagingOperationValue.Receive") {
+    val ours = MessagingOperationTypeValue.Receive.value
+    val otel = otel4sMessaging.MessagingOperationTypeValue.Receive.value
+    assertEquals(ours, otel)
+  }
+
+  test("MessagingMessageId") {
+    val value = "value"
+    val ours = MessagingMessageId(value)
+    val otel = otel4sMessaging.MessagingMessageId(value)
+    assertEquals(ours, otel)
+  }
+
+  test("MessagingDestinationName") {
+    val value = "value"
+    val ours = MessagingDestinationName(value)
+    val otel = otel4sMessaging.MessagingDestinationName(value)
+    assertEquals(ours, otel)
+  }
+
 }
